[REFACTOR]: 테스트를 위한 RedisService 인터페이스화
본문

[#?]
