+++ src/main/java/project/common/service/RedisService.java
@@ -1,34 +1,12 @@
 package project.common.service;
 
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.data.redis.core.ValueOperations;
-import java.util.concurrent.TimeUnit;
+public interface RedisService {
 
-@Service
-@RequiredArgsConstructor
-public class RedisService{
-    
-    private final RedisTemplate redisTemplate;
-    
-    public String getValues(String key){
-        ValueOperations<String, String> values = redisTemplate.opsForValue();
-        return values.get(key);
-    }
-    
-    public void setValues(String key, String value){
-        ValueOperations<String, String> values = redisTemplate.opsForValue();
-        values.set(key,value);
-    }
-    
-    public void setExpiration(String key, Long time){
-        int milliSeconds = time.intValue();
-        
-        redisTemplate.expire(key, milliSeconds, TimeUnit.MILLISECONDS);
-    }
-    
-    public void deleteByKey(String key){
-        redisTemplate.delete(key);
-    }
+    public String getValues(String key);
+
+    public void setValues(String key, String value);
+
+    public void setExpiration(String key, Long time);
+
+    public void deleteByKey(String key);
 }
